correction in time and more frequent removal of temp files

+++ src/dataset/downloads/make_dataset.py
@@ -504,15 +504,16 @@ for link in link_list:
 	overpass.randomize_boxes_offset()
 	overpass.process_boxes()
 
-# Remove temporary files
-if (used_remove == True):	
-	dir_path = Path(storage_path_temp) 
-	try:
-	    shutil.rmtree(dir_path)
-	except OSError as e:
-	    print("Error: %s : %s" % (dir_path, e.strerror))
+	if(i%10==0):
+		# Remove temporary files
+		if (used_remove == True):	
+			dir_path = Path(storage_path_temp) 
+			try:
+			    shutil.rmtree(dir_path)
+			except OSError as e:
+			    print("Error: %s : %s" % (dir_path, e.strerror))
+						
 				
-
